requirements.txt
neo4j==5.15.0
spacy==3.8.7
sentence-transformers==2.7.0
networkx==3.5
scikit-learn==1.7.0
numpy==1.26.4
pandas==2.3.0
PyPDF2==3.0.1
python-dotenv==1.1.1
openai==1.92.2
tiktoken==0.9.0
graphrag==2.3.0
reportlab==4.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles>=24.1.0,<25.0.0
nest-asyncio==1.6.0
python-louvain==0.16
email-validator==2.2.0
psycopg2-binary==2.9.9
spacy
sentence-transformers
numpy
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl